vtgate/buffer: Refactored failover error detection code.
Removed open-source specific file and merged the detection for all known flavors into one function instead.
